requirements.txt
# Core ML & Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0
transformers>=4.35.0
torch-geometric>=2.4.0
networkx>=3.0
node2vec>=0.4.6

# MLOps & Tracking
mlflow>=2.8.0
optuna>=3.4.0

# Data Processing & Streaming
pyarrow>=14.0.0
kafka-python>=2.0.2
pyspark>=3.5.0

# Web Frameworks
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
gunicorn>=21.2.0
grpcio>=1.59.0
grpcio-tools>=1.59.0

# Database & Storage
boto3>=1.29.0
psycopg2-binary>=2.9.9
clickhouse-driver>=0.2.6
elasticsearch>=8.11.0
redis>=5.0.0

# Security & Crypto
cryptography>=41.0.0
pyjwt>=2.8.0
python-multipart>=0.0.6

# Computer Vision & Face Recognition
opencv-python>=4.8.0
dlib>=19.24.0
face-recognition>=1.3.0
pillow>=10.0.0

# Monitoring & Observability
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Explainability
shap>=0.43.0
captum>=0.6.0

# Email & Network Analysis
dnspython>=2.4.2
validators>=0.22.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3

# PE/Binary Analysis
pefile>=2023.2.7
ssdeep>=3.4
yara-python>=4.3.1

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.7
tqdm>=4.66.0
joblib>=1.3.2

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
bandit>=1.7.5

# Development
ipython>=8.17.0
jupyter>=1.0.0
notebook>=7.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
